arm64: IOP (R3000A) recompiler
ARM64 IOP dynarec and opcode tables, plus the s32 cycle-delta cast fix (host FCVTZS
saturation differs from x86 CVTTSS2SI on out-of-range inputs).

+++ pcsx2/IopBios.cpp
@@ -854,7 +854,7 @@ namespace R3000A
 				v0 = file->read(buf.get(), count);
 
 				[[likely]]
-				if (v0 >= 0 && iopMemSafeWriteBytes(data, buf.get(), v0))
+				if (static_cast<s32>(v0) >= 0 && iopMemSafeWriteBytes(data, buf.get(), v0))
 				{
 					psxCpu->Clear(data, (v0 + 3) / 4);
 				}
